Including another option for view configuration icons
Since the minimalist icons for various view configurations in the ViewerConfiguration subdirs for both themes look so self-similar, it's hard to visually parse them.  I added a little more color coding for 3D and Slice viewers to make them simpler to tell apart and understand. We should let developers decide which is preferred and keep only that one.